requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
google-genai
loguru
google-search-results
gh-md-to-html
pdfkit
xhtml2pdf
jinja2
aiofiles
python-multipart
pydantic[email]
lxml_html_clean
pydantic[email]
gh-md-to-html[pdf_export]
xhtml2pdf
